feat: add controllers and updates

+++ src/main/java/com/foro/domain/service/CursoService.java
@@ -1,6 +1,7 @@
 package com.foro.domain.service;
 
 import com.foro.domain.entity.CursoEntity;
+import com.foro.exception.NoContentException;
 import com.foro.exception.NotFoundException;
 import com.foro.persistence.dto.CursoDto;
 import com.foro.persistence.mapper.CursoMapper;
@@ -9,7 +10,6 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.NoSuchElementException;
 import java.util.Optional;
 
 @Service
@@ -32,8 +32,14 @@ public class CursoService {
         return cursoRepository.existsById(id);
     }
 
-    public List<CursoDto> findAll() {
-        return cursoMapper.toCursoList(cursoRepository.findAll());
+    public List<CursoDto> findAll() throws NoContentException {
+        List<CursoEntity> cursos = cursoRepository.findAll();
+
+        if (cursos.isEmpty()) {
+            throw new NoContentException();
+        }
+
+        return cursoMapper.toCursoList(cursos);
     }
 
     public CursoDto update(CursoDto cursoDto) throws NotFoundException {
@@ -50,7 +56,11 @@ public class CursoService {
         }).orElseThrow(NotFoundException::new);
     }
 
-    public void deleteById(Long id) {
-        cursoRepository.deleteById(id);
+    public void deleteById(Long id) throws NotFoundException {
+        if (existsById(id)) {
+            cursoRepository.deleteById(id);
+        } else {
+            throw new NotFoundException();
+        }
     }
 }
